requirements.txt
imutils==0.5.4
keras==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.7.1
mpmath==1.2.1
numpy==1.24.3
pandas==2.0.1
parse==1.19.0
piexif==1.1.3
Pillow==9.5.0
scikit-learn==1.2.2
scipy==1.10.1
tensorflow==2.10.0
torch==2.0.1+cu118
tqdm==4.65.0